ai-server/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
gotrue==2.6.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.40.1
packaging==24.1
playwright==1.45.1
postgrest==0.16.9
pydantic==2.8.2
pydantic-core==2.20.1
pyee==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.2
watchdog==4.0.1
websockets==12.0
werkzeug==3.0.3
zipp==3.19.2
